requirements.txt
asgiref==3.5.2
cloudinary==1.29.0
dj-database-url==1.0.0
dj3-cloudinary-storage==0.0.6
Django==3.2.15
django-bootstrap-datepicker-plus==4.0.0
django-bootstrap-v5==1.0.11
django-bootstrap4==22.2
django-phone-field==1.8.1
django-phonenumber-field==7.0.0
gunicorn==20.1.0
phonenumbers==8.12.55
psycopg2==2.9.3
pytz==2022.2.1
sqlparse==0.4.2
